refactor(core): move option types to owning modules
Apply "Types Follow Functions" principle:
- HttpOptions, ControllerClass defined in http-module.ts
- SchedulerClass defined in scheduler-module.ts
- Remove internal types from public API exports

Users only need CreateAppOptions; internal details are hidden.

Verified: precommit-verify ✓ (0728fdfedec1823d)
