+++ packages/core/src/app/index.ts
@@ -4,9 +4,6 @@ export type {
   HttpApp,
   CommandApp,
   CreateAppOptions,
-  HttpOptions,
   ReadyOptions,
   ReadyResult,
-  ControllerClass,
-  SchedulerClass,
 } from './types';
